java-sv2-daily-labs09/week09-lab-tasks/src/main/java/day01/FrontOfStageTicket.java
package day01;

import java.time.LocalDateTime;
import java.time.LocalTime;

public class FrontOfStageTicket extends Ticket {

    private String extraCode;

    public FrontOfStageTicket(String bandName, LocalDateTime dateTime, int price, String extraCode) {
        super(bandName, dateTime, price);
        this.extraCode = extraCode;
    }

    @Override
    public LocalTime entryTime() {
        return super.entryTime().minusHours(1);
    }

    public String getExtraCode() {
        return extraCode;
    }
}
